+++ src/main/java/org/gestern/gringotts/TransactionResult.java
@@ -1,0 +1,10 @@
+package org.gestern.gringotts;
+
+/**
+ * The result of a transaction.
+ * @author jast
+ *
+ */
+public enum TransactionResult {
+	SUCCESS, INSUFFICIENT_FUNDS, INSUFFICIENT_SPACE, ERROR;
+}